requirements.txt
transformers==3.0.0
wikipedia
pipeline
nltk